+++ src/Constraints/ViewExists.php
@@ -1,19 +1,13 @@
 <?php
 
-namespace Sven\PackageTestingUtils\Constraints;
+namespace Sven\LaravelTestingUtils\Constraints;
 
 use Illuminate\Contracts\View\Factory;
 use PHPUnit\Framework\Constraint\Constraint;
 
-/**
- * Class ViewExists
- */
 class ViewExists extends Constraint
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function matches($other)
+    public function matches($other): bool
     {
         /** @var \Illuminate\View\Factory $viewFactory */
         $viewFactory = app(Factory::class);
@@ -21,12 +15,7 @@ class ViewExists extends Constraint
         return $viewFactory->exists($other);
     }
 
-    /**
-     * Returns a string representation of the object.
-     *
-     * @return string
-     */
-    public function toString()
+    public function toString(): string
     {
         return 'does exist';
     }